+++ viadot/sources/__init__.py
@@ -9,7 +9,7 @@ from .sharepoint import Sharepoint
 from .supermetrics import Supermetrics
 from .genesys import Genesys
 from .sftp import SftpConnector
-
+from .mediatool import Mediatool
 
 try:
     from .sap_rfc import SAPRFC