+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `validate_date_filter` parameter to `Epicor` source, `EpicorOrdersToDF` task and `EpicorOrdersToDuckDB` flow.
 This parameter enables user to decide whether or not filter should be validated.
+- Added `Mediatool` source class
+- Added `MediatoolToDF` task class
+- Added `MediatoolToADLS` flow class
+- Added option to disable `check_dtypes_sort` in class/flow_name.
+
 
 ## [0.4.12] - 2023-01-31
 ### Added
@@ -18,7 +23,6 @@ This parameter enables user to decide whether or not filter should be validated.
 - Added `timeout` parameter to all `Flow`s where it can be added.
 - Added `adls_bulk_upload` task function to `task_utils.py`
 - Added `get_survey_list` into `Mindful` Source file.
-
 
 ### Changed
 - Updated `genesys_to_adls.py` flow with the `adls_bulk_upload` task